packages/mcp-server/pyproject.toml
[project]
name = "mame-mcp"
version = "0.1.0"
description = "MCP server — Materials Project API bridge for MAME"
requires-python = ">=3.11"
dependencies = [
    "mp-api>=0.46.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mame_mcp"]

[project.scripts]
mame-mcp = "mame_mcp.server:main"
